XML files for a Civ VI mod
These are XML files for a mod I am trying to work on for Civilizations 6. I was unable to make the files work so I have moved on to transferring the data on these files to an SQL template by Keniisu and have had more luck running the mod. Still transferring data so fingers crossed that once it is all transferred, things will work.